requirements.txt
SQLAlchemy==1.3.13
pandas==0.25.3
numpy==1.18.1
selenium==3.141.0
python-dotenv==0.12.0
